+++ src/components/dialogs/commons/basicModificationDialog.js
@@ -9,7 +9,7 @@ import React from 'react';
 import { FormattedMessage } from 'react-intl';
 import { Button } from '@mui/material';
 import PropTypes from 'prop-types';
-import ModificationDialogCommon from './modification-dialog-common';
+import ModificationDialogContent from './modification-dialog-content';
 
 /**
  * Generic Modification Dialog which manage basic common behaviors
@@ -44,7 +44,7 @@ const BasicModificationDialog = ({
     );
 
     return (
-        <ModificationDialogCommon
+        <ModificationDialogContent
             submitButton={submitButton}
             closeAndClear={closeAndClear}
             {...props}
